TG-91: fix Profile component, delete hooks folder
